% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_network.R
\name{generate_network}
\alias{generate_network}
\title{Generate Network}
\usage{
generate_network(net_name, net_params)
}
\arguments{
\item{net_name}{String, the name of netwrok}

\item{net_params}{Vector, the params of generating network}
}
\value{
Network, List, store the name and graph generated by igraph
}
\description{
Generate 3 kinds of networks for simulating. You can use
Random Network, Small World Network and Scale-Free Network.
}
\examples{
generate_network("Random Network", c(10, "gnm", 10, "TRUE"))
generate_network("Random Network", c(10, 0.2))
}
