% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gif.R
\name{plot_gif}
\alias{plot_gif}
\title{Plot GIF}
\usage{
plot_gif(Var_Param, layout.old)
}
\arguments{
\item{Var_Param}{List, saved important var and params}

\item{layout.old}{Grid Fruchterman-Reingold layout}
}
\description{
Plot each frame of GIF with grid layout.old.
}
\examples{
Var_Param <- set_param("Random Network", c(0.2), 10, 70, 1, 0, 0.12, 0.4, 7)
layout.old = igraph::layout.fruchterman.reingold(Var_Param$Network$graph, niter = 1000)
plot_gif(Var_Param, layout.old)
}
