% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_selection.R
\name{random_selection}
\alias{random_selection}
\title{Randomly decide whether this node be infected of removed}
\usage{
random_selection(freq, vec_porb)
}
\arguments{
\item{freq}{int, the number of incoming links(removed: freq == 1)}

\item{vec_porb}{vector, the probility vector of vec_rate and 1-vec_rate}
}
\value{
results, int
}
\description{
Infection: Randomly decide whether it will infect the target node
use sample(c(1, 0), 1, prob = ) for every incoming link.If it will
be infected then return 1, otherwise return 0.Finally, add the
return value of each incoming link to find out if this node will
be infected.
Removed: Randomly decide whether the target node will be removed
from infected queue. This is a special condition with freq == 1
}
\examples{
random_selection(4, c(0.12, 0.88))
}
