% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gif.R
\name{save_gif}
\alias{save_gif}
\title{Save GIF}
\usage{
save_gif(Var_Param, layout.old)
}
\arguments{
\item{Var_Param}{List, saved important var and params}

\item{layout.old}{Grid Fruchterman-Reingold layout}
}
\value{
GIF Image
}
\description{
Generate GIF image from frames.
}
\examples{
Var_Param <- set_param("Random Network", c(0.2), 10, 70, 1, 0, 0.12, 0.4, 7)
layout.old = igraph::layout.fruchterman.reingold(Var_Param$Network$graph, niter = 1000)
save_gif(Var_Param, layout.old)
}
