% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_param.R
\name{set_param}
\alias{set_param}
\title{Set important params}
\usage{
set_param(
  Net_Name,
  Net_Params,
  N,
  TIME,
  I,
  R,
  InfectedRate,
  RemovedRate,
  Latency_Infected
)
}
\arguments{
\item{Net_Name}{String, 网络名称}

\item{Net_Params}{Vector, 网络参数向量}

\item{N}{Int, 模拟节点数}

\item{TIME}{Int, 模拟总时间}

\item{I}{Int, 初始感染人数}

\item{R}{Int, 初始移除人数}

\item{InfectedRate}{Float, 感染率}

\item{RemovedRate}{Float, 移除率}

\item{Latency_Infected}{Int, 潜伏期}
}
\value{
Var_Param List
}
\description{
Set important params of simulation and create a list as
a structure in C to pass the important variables and params.
It contains:
N Total number of nodes
TIME Total time of simulation
S I R Imitial number of S、I、R
InfectedRate Infected rate
RemovedRate Removed rate
Latency_Infected Latency of infected
TotalInfected The total number of infected
Infected_list Store the number of infected changeing over time
Removed_list Store the number of removed changeing over time
TimeLine_Infected Store the latency over time for every node
TimeLine_Removed
}
\examples{
set_param("Random Network", c(0.2), 10, 70, 1, 0, 0.12, 0.4, 7)
}
