% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate}
\usage{
simulate(Var_Param)
}
\arguments{
\item{Var_Param}{List, saved important var and params}
}
\value{
Show the GIF
}
\description{
Simulate the process.
}
\examples{
Var_Param <- set_param("Random Network", c(0.2), 10, 70, 1, 0, 0.12, 0.4, 7)
simulate(Var_Param)
}
