% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_infected.R
\name{update_infected}
\alias{update_infected}
\title{Update infected}
\usage{
update_infected(Var_Param, total_time)
}
\arguments{
\item{Var_Param}{List, saved important var and params}

\item{total_time}{Int, the time at now}
}
\value{
infected, Vector, the new infected nodes at now
}
\description{
Randomly update newly infected nodes and return all the infected nodes.
}
\examples{
Var_Param <- set_param("Random Network", c(0.2), 10, 70, 1, 0, 0.12, 0.4, 7)
infected <- sample(igraph::V(Var_Param$Network$graph), Var_Param$I)
Var_Param$Infected_list[[1]] = infected
for (i in 1:Var_Param$Latency_Infected) {
Var_Param$TimeLine_Infected[[i]] = igraph::V(Var_Param$Network$graph)[0]
}
Var_Param$TimeLine_Infected[[Var_Param$Latency_Infected]] = infected
Var_Param$Removed_list[[1]] = igraph::V(Var_Param$Network$graph)[0]
update_infected(Var_Param, 1)
}
