% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_removed.R
\name{update_removed}
\alias{update_removed}
\title{Update removed}
\usage{
update_removed(Var_Param, wait_removed, day)
}
\arguments{
\item{Var_Param}{List, saved important var and params}

\item{wait_removed}{Vector, the infected nodes wait to be removed}

\item{day}{Int, the time at now}
}
\value{
Var_Param$Removed_list List, shallow copy
}
\description{
Randomly update removed infected nodes and return all the removed nodes.
}
\examples{
NULL
}
